+++ src/test/java/activation/ReluActivationTest.java
@@ -1,8 +1,6 @@
 package activation;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.provider.Arguments;
-
 import java.util.stream.Stream;
 
 import static utils.TestGenerator.initDefinedNeuronVector;
@@ -18,14 +16,14 @@ public class ReluActivationTest extends OutputActivationFuncTest
     protected Stream<Arguments> provideUpdateToVectorXTestData() {
         return Stream.of(
                 Arguments.of("Negative input",
-                        initDefinedNeuronVector(-42.1, -5.1584),
-                        new double[]{0.0, 0.0}),
+                        initDefinedNeuronVector(-42.1f, -5.1584f),
+                        new float[]{0.0f, 0.0f}),
                 Arguments.of("Positive input",
-                        initDefinedNeuronVector(3.215, 13.421),
-                        new double[]{3.215, 13.421}),
+                        initDefinedNeuronVector(3.215f, 13.421f),
+                        new float[]{3.215f, 13.421f}),
                 Arguments.of("Mixed input",
-                        initDefinedNeuronVector(-4.391, 0.0, 5.952, 9.2, -9.291521),
-                        new double[]{0.0, 0.0, 5.952, 9.2, 0.0})
+                        initDefinedNeuronVector(-4.391f, 0.0f, 5.952f, 9.2f, -9.291521f),
+                        new float[]{0.0f, 0.0f, 5.952f, 9.2f, 0.0f})
         );
     };
 }
