+++ src/test/java/loss/LossFuncTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import structure.Neuron;
-import structure.Scalar;
 import utils.TestGenerator;
 
 import java.util.stream.Stream;
@@ -22,23 +21,23 @@ public abstract class LossFuncTest
     @Test
     public void shouldReturnPositiveCost_whenComputeIsUsed() {
         //given
-        Neuron[] predicted = TestGenerator.initRandomNeuronVector(20, 1.0);
-        double [] target = TestGenerator.generateRandomVector(20, 1.0);
+        Neuron[] predicted = TestGenerator.initRandomNeuronVector(20, 1.0f);
+        float [] target = TestGenerator.generateRandomVector(20, 1.0f);
 
         //when
-        double result = getLossFunc().compute(predicted, target);
+        float result = getLossFunc().compute(predicted, target);
 
         //then
-        assertTrue(result > 0.0);
+        assertTrue(result > 0.0f);
     }
 
     @ParameterizedTest(name = "{0}")
     @MethodSource("provideReturnXCostTestData")
-    public void shouldReturnXCost_whenComputeIsUsed(String description, Neuron [] predicted, double [] target, double expected, double delta) {
+    public void shouldReturnXCost_whenComputeIsUsed(String description, Neuron [] predicted, float [] target, float expected, float delta) {
         //given
 
         //when
-        double result = getLossFunc().compute(predicted, target);
+        float result = getLossFunc().compute(predicted, target);
 
         //then
         assertEquals(expected, result, delta);
@@ -47,13 +46,10 @@ public abstract class LossFuncTest
     @Test
     public void shouldThrowException_whenInputSizesAreDifferent() {
         //given
-        Neuron [] predicted = TestGenerator.initRandomNeuronVector(5, 1.2);
-        double [] target = TestGenerator.generateRandomVector(6, 1.2);
+        Neuron [] predicted = TestGenerator.initRandomNeuronVector(5, 1.2f);
+        float [] target = TestGenerator.generateRandomVector(6, 1.2f);
 
         //then
         assertThrows(IllegalArgumentException.class, () -> getLossFunc().compute(predicted, target));
     }
-
-    //derive tests
-
 }
